# [GLUTEN-3668][CH]Fix to_date function performance