## What changes were proposed in this pull request?

(Please fill in changes proposed in this fix)

(Fixes: \#3668)

## How was this patch tested?

TEST BY UT

性能测试数据


3000W 行正常数据 
测试SQL: `select count(1) from $test_tbl where to_date($col) > '1990-01-01'`
PR改动前耗时： 2.983s， 2.686s,   2.804s
PR改动后耗时： 2.94s，2.861s，2.842s；

3000W行数据 (其中2500W行是NULL，500W是正常数据)
测试SQL:  `select count(1) from $test_tbl where to_date($col) > '1990-01-01'`
PR改动前耗时：0.621s， 0.614s,  0.677s
PR改动后耗时：0.631s，0.641s，0.692s；

3000W行数据 (其中2500W数据是不符合日期格式的随机字符串，500W行是正常数据) 
测试SQL: `select count(1) from $test_tbl where to_date($col) > '1990-01-01'`
PR改动前耗时：6.148s，6.018s，5.845s
PR改动后耗时：3.188s，3.055s，3.08s

对比发现，正常数据测试情况下性能接近，在某些异常场景下性能有所提升