@KevinyhZou 2023-11-15
> to_date 是支持指定日期格式，这个有考虑了？

我考虑下这个