@KevinyhZou 2023-11-16
> to_date 是支持指定日期格式，这个有考虑了？

to_date 指定日期格式不需要考虑支持，如果使用to_date($col, 'yyyy-MM-dd') 按照原本的指定，会使用parseDateTimeInJodaSyntaxOrNull 来 解析